Also print timezone
This adresses issue #2 a bit.

One of my machine was way off the time normal,
and also was sete to UTC (which is fine to me),
but this was not reflected on the timestamp of the login,
so it was a bit tricky to synchronize the phone
to the correct time to get a proper TOTP from it.